+++ src/users/mod.rs
@@ -1,2 +1,3 @@
+pub mod account_service;
 pub mod auth_service;
 pub mod permission;